Switching to .ts/.tsx
