+++ jest.config.js
@@ -1,0 +1,10 @@
+const jestConfig = {
+    testEnvironment: 'jsdom',
+    setupFilesAfterEnv: ['<rootDir>/src/setupTests.js'],
+    moduleNameMapper: {
+      '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
+    },
+  }
+  
+  export default jestConfig
+
