+++ package.json
@@ -12,9 +12,8 @@
     "@headlessui/react": "^2.1.9",
     "@heroicons/react": "^1.0.6",
     "@hookform/resolvers": "^3.3.4",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
     "axios": "^1.6.8",
     "bootstrap": "^5.3.3",
     "docx": "^9.0.2",
@@ -32,7 +31,6 @@
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.51.4",
     "react-router-dom": "^6.23.0",
-    "react-scripts": "^5.0.1",
     "react-table": "^7.8.0",
     "react-toastify": "^10.0.5",
     "react-tooltip": "^5.28.0",
@@ -44,11 +42,11 @@
     "yup": "^1.4.0"
   },
   "scripts": {
-    "start": "react-scripts start --openssl-legacy-provider",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "dev": "react-scripts start"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest run",
+    "test:watch": "vitest"
   },
   "eslintConfig": {
     "extends": [
@@ -70,6 +68,13 @@
   },
   "devDependencies": {
     "@babel/plugin-transform-json-strings": "^7.18.6",
-    "@babel/plugin-transform-optional-catch-binding": "^7.18.6"
+    "@babel/plugin-transform-optional-catch-binding": "^7.18.6",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@vitejs/plugin-react": "^4.4.1",
+    "jsdom": "^26.1.0",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.2",
+    "vitest": "^3.1.2"
   }
 }
