+++ src/components/Contract/ContractDetails/ContractDetails.tsx
