+++ src/setupTests.ts
@@ -1,0 +1,4 @@
+import { expect } from 'vitest' // ✅ Vitest's own expect
+import * as matchers from '@testing-library/jest-dom/matchers'
+
+expect.extend(matchers)
